little file renaming
