+++ frontend/webpack-stats.json
@@ -1,1 +1,1 @@
-{"status":"done","publicPath":"http://0.0.0.0:8080/","chunks":{"app":[{"name":"bundle.css","publicPath":"http://0.0.0.0:8080/bundle.css","path":"/home/lorenzo/Documenti/Other/evs/frontend/dist/bundle.css"},{"name":"bundle.js","publicPath":"http://0.0.0.0:8080/bundle.js","path":"/home/lorenzo/Documenti/Other/evs/frontend/dist/bundle.js"},{"name":"app.f12d738c8ec08f4f2626.hot-update.js","publicPath":"http://0.0.0.0:8080/app.f12d738c8ec08f4f2626.hot-update.js","path":"/home/lorenzo/Documenti/Other/evs/frontend/dist/app.f12d738c8ec08f4f2626.hot-update.js"}]},"error":"ModuleError","message":"Module Error (from ./node_modules/eslint-loader/index.js):\n\n/home/lorenzo/Documenti/Other/evs/frontend/src/components/events/EventForm.vue\n  99:15  error  'v-bind' directives require an attribute value  vue/valid-v-bind\n\n✖ 1 problem (1 error, 0 warnings)\n"}
+{"status":"done","publicPath":"http://0.0.0.0:8080/","chunks":{"app":[{"name":"bundle.css","publicPath":"http://0.0.0.0:8080/bundle.css","path":"/home/lorenzo/Documenti/Other/evs/frontend/dist/bundle.css"},{"name":"bundle.js","publicPath":"http://0.0.0.0:8080/bundle.js","path":"/home/lorenzo/Documenti/Other/evs/frontend/dist/bundle.js"},{"name":"app.ee98ce80fb19c73ae543.hot-update.js","publicPath":"http://0.0.0.0:8080/app.ee98ce80fb19c73ae543.hot-update.js","path":"/home/lorenzo/Documenti/Other/evs/frontend/dist/app.ee98ce80fb19c73ae543.hot-update.js"}]}}
